requirements-checkpoint.txt
streamlit
langchain
sentence-transformers
faiss-cpu
gitpython
transformers
tree-sitter